BUG: Use launcher for local Python pre-commit hooks
Python may not be available on Windows systems. Use a launcher script to
run local Python pre-commit hooks with the Python executable we generate
when bootstrapping our pre-commit hooks. The launcher script uses the
local Python, which we know will be available and have a sufficient
version. It also allows adjusting the path with Windows vs Unix.