BUG: Do not pass filename to local hooks
When a large number of files are commit the local pre-commit hooks
were being called multiple times with different filename
arguments. Fix by disabling passing filenames to ensure that the hook
is only called onces. These hooks inspect the state of git to
determine what changes are staged for commit.