COMP: Remove LEGACY off mode in backport
The backported AnatomicalOrientation is explicitly available
for the purpose of providing an easy long-term conversion
to the new paradigm.

Have conditional support for deprecating features of future
support makes the backport unnecessarily complicated.

Using a separate commit to remove the backported future
deprecation elements to minimize the changes from the original
files.