BUG: MINCImageIO should use `signed char` for `CHAR` and `MI_TYPE_BYTE`
`MINCImageIO::Write` should use `signed char` for the estimation of the min and
the max pixel value, when the pixel component type is `IOComponentEnum::CHAR`.
Both `IOComponentEnum::CHAR` and `MI_TYPE_BYTE` specify a _signed_ type, whereas
the default `char` might be unsigned (for example, when using GCC compiler
option `-funsigned-char`).

Obviously the estimation of min and max pixel value may go wrong when the pixels
are incorrectly assumed to be unsigned.