DOC: Fix all the spelling bugs
Now that the spell checker is being run on all the CXX files
in ITK, we need to fix some of the spell bugs that existed.

Corrected various typos and inconsistencies in code and documentation,
such as replacing "geom" with "geometry," and allowing US and English version of
terms like "initialized/initialised," "optimization/opimisation".

Change test, localscope variagbles like "transformIO" for better clarity and consistency.

Co-authored-by: Hans Johnson <hans-johnson@uiowa.edu>