ENH: Spell checking pre-commit
This PR address a spell checking pre-commit to ITK. So any time code is
committed to an ITK repo, the SimpleITK/CommentSpellCheck (CSC) script is
run on the new code. CSC checks the spelling of all comments in new file.

Also the version of CSC use is upgraded to 0.4.3, which uses a pure Python
spelling backend, unlike the previous version. That previous version depended
on the C enchant library, which was not available on all systems.

We have removed the spell checking github action, since it is no longer
required.  Any time a pull request is submitted to github, pre-commit
is run.