STYLE: Remove trailing `std::endl` inserts from itkExceptionMacro calls
It's uncommon and unnecessary to have an extra newline and flush at the end of
an exception message. It is the responsibility of the one who prints the
exception description to add a line break, if necessary (like
`ExceptionObject::Print` does).

Note also that these `std::endl` manipulators were inserted into an
`std::ostringstream`, and for `std::ostringstream`, flushing does not have any
effect anyway.

Using Notepad++, Replace In Files:

    Find what: (  itkExceptionMacro\([^)]+) << std::endl(\);)
    Find what: (  itkExceptionMacro\([^;]+) << std::endl(\);)
    Replace with: $1$2
    Directory: D:\ITK\Modules\
    [v] Match case
    (*) Regular expression [v] . matches newline