ENH: Minimize differences with upstream expat code
ITK version of expat was last fully syncronized with
expat hash 72d7ce953827fe08a56b8ea64092f208be6ffc5b from
2021-09-20.

In that version of the code the XMLPARSEAPI macros
are defined in expat_external.h (which is unconditionally
included in expat.h).

Move library source files to lib subdirectory to match
the organization in the libexpat upstream.