ENH: Update to expat 2.7.2 code changes
Manually upgrade from expat 2.4.1 to 2.7.2.

EXPAT_GE and XML_GE need to be set in expat 2.7.2, but do not
need to be viewed in ITK.

If the variable EXPAT_SHARED_LIBS is true, then we
must set the value of ITK_EXPAT_STATIC to 1, otherwise
the variable ITK_EXPAT_STATIC is expected to not exist.

Minimize the code differences between ITK and expat upstream 2.7.2 with
regards to dllexport and dllimport.

Minimize the code differences between ITK and expat upstream 2.7.2 with
regards to dllexport and dllimport.

Remove unused expatConfig.h configuration file

Remove unused expatDllConfig.h configuration file

Remove dependence on libexpat.def MSVC file

By using the upstream configuration for dllimport and dll export, it is no longer necessary to
create a compile time file to determine if static or shared libraries are used.

On linux build use the following to identify symbols in 2.7.2
```bash
nm lib/libITKEXPAT-6.0.a \
     | grep " [TRD] " \
     |  awk '{ print $3 }' \
     | sort -df \
     | awk '{ print "#define "$1" @MANGLE_PREFIX@_"$1 }'
```