BUG: `EXPECT_EQ` in `while` loop in BSplineTransformGTest unreachable
The condition `!iter1.IsAtEnd() && iter2.IsAtEnd()` is usually false, because
both iterators are expected to iterate over the same (buffered) region.

It should be sufficient to just check if the first iterator is not at the end.
And then, it should be safe to _assert_ that the second iterator is not at the
end either.